+++ src/components/Header/Header.tsx
@@ -2,7 +2,7 @@
 
 import { type ReactElement } from 'react';
 
-import { TEXT_CONTENT } from '@/constants/TEXT_CONTENT';
+import { TEXT_CONTENT } from '@/constants/const-text-content';
 
 const Header = (): ReactElement => {
   return (
