+++ src/components/Timer/Timer.tsx
@@ -6,10 +6,10 @@ import { useEffect, useState, type ReactElement } from 'react';
 import { useSelector } from 'react-redux';
 import { useDispatch } from 'react-redux';
 
-import { TIMER_CONFIG } from '@/constants/TIMER_CONFIG';
+import { TIMER_CONFIG } from '@/constants/const-timer-config';
 import { endTestFailure, setWrapperOpacity } from '@/redux/appStateSlice';
 import type { RootState } from '@/redux/store';
-import { localStorageUtil } from '@/utils/localStorage';
+import { localStorageUtil } from '@/utils/localStorageUtil';
 
 const Timer = (): ReactElement => {
   const dispatch = useDispatch();
