+++ src/constants/const-timer-config.ts
